Factor out LastBackStorage to separate file
